Build rand without default features (no_std)
The tests using rand/std have been moved to ci/big_rand/.
