+++ .travis.yml
@@ -26,7 +26,7 @@ matrix:
       before_script:
         - rustup target add $TARGET
       script:
-        - cargo build --verbose --target $TARGET --no-default-features --features "serde"
+        - cargo build --verbose --target $TARGET --no-default-features --features "serde rand"
     - name: "rustfmt"
       rust: 1.31.0
       before_script:
