+++ tests/bigint.rs
@@ -1,8 +1,6 @@
 extern crate num_bigint;
 extern crate num_integer;
 extern crate num_traits;
-#[cfg(feature = "rand")]
-extern crate rand;
 
 use num_bigint::BigUint;
 use num_bigint::Sign::{Minus, NoSign, Plus};
@@ -1081,25 +1079,6 @@ fn test_negative_shr() {
     assert_eq!(BigInt::from(-2) >> 1, BigInt::from(-1));
     assert_eq!(BigInt::from(-3) >> 1, BigInt::from(-2));
     assert_eq!(BigInt::from(-3) >> 2, BigInt::from(-1));
-}
-
-#[test]
-#[cfg(feature = "rand")]
-fn test_random_shr() {
-    use rand::distributions::Standard;
-    use rand::Rng;
-    let rng = rand::thread_rng();
-
-    for p in rng.sample_iter::<i64, _>(&Standard).take(1000) {
-        let big = BigInt::from(p);
-        let bigger = &big << 1000;
-        assert_eq!(&bigger >> 1000, big);
-        for i in 0..64 {
-            let answer = BigInt::from(p >> i);
-            assert_eq!(&big >> i, answer);
-            assert_eq!(&bigger >> (1000 + i), answer);
-        }
-    }
 }
 
 #[test]
